break out dupe code into a helper for integration tests and system tests to share
