+++ test/test_helper.rb
@@ -11,6 +11,7 @@ require 'minitest/autorun'
 require 'capybara/rails'
 require 'capybara-screenshot/minitest'
 require 'omniauth_helper'
+require 'integration_helper'
 require 'rack/test'
 
 if ENV['CIRCLE_ARTIFACTS']
@@ -49,7 +50,7 @@ class ActiveSupport::TestCase
   end
 
   def create_language_config
-    language_options = %w(Spanish French Korean)
+    language_options = %w[Spanish French Korean]
     create :config, config_key: 'language',
                     config_value: { options: language_options }
   end
@@ -66,6 +67,7 @@ end
 class ActionDispatch::IntegrationTest
   include Capybara::DSL
   include Capybara::Screenshot::MiniTestPlugin
+  include IntegrationHelper
   include OmniauthMocker
   OmniAuth.config.test_mode = true
 
@@ -80,56 +82,5 @@ class ActionDispatch::IntegrationTest
 
   def choose_line(line)
     post lines_path, params: { line: line.to_s }
-  end
-
-  # for proper integration tests
-  def log_in_as(user, line = 'DC')
-    log_in user
-    select_line line
-  end
-
-  def log_in(user)
-    visit root_path
-    fill_in 'Email', with: user.email
-    fill_in 'Password', with: user.password
-    click_button 'Sign in with password'
-  end
-
-  def select_line(line = 'DC')
-    choose line
-    click_button 'Start'
-  end
-
-  def wait_for_element(text)
-    has_content? text
-  end
-
-  def wait_for_no_element(text)
-    has_no_content? text
-  end
-
-  def sign_out
-    click_link @user.name
-    click_link 'Sign Out'
-  end
-
-  def wait_for_ajax
-    Timeout.timeout(Capybara.default_max_wait_time) do
-      loop until _finished_all_ajax_requests?
-    end
-  end
-
-  def _finished_all_ajax_requests?
-    page.evaluate_script('jQuery.active').zero?
-  end
-
-  def go_to_dashboard
-    click_link "DARIA - #{(ENV['FUND'] ? ENV['FUND'] : Rails.env)}"
-  end
-
-  def click_away_from_field
-    find('body').click
-    fill_in 'First and last name', with: nil
-    wait_for_ajax
   end
 end
